fix: split product dashboard into server wrapper\n\n- move product dashboard CSR logic into ProductDashboardClient\n- wrap page.tsx in Suspense to satisfy useSearchParams\n- wire up ClientChatWidget helper for chat widget lazy load\n\nDroid-assisted
