+++ frontend/src/components/ClientChatWidget.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import dynamic from 'next/dynamic';
+
+const ChatWidget = dynamic(() => import('./ChatWidget'), { ssr: false });
+
+export default function ClientChatWidget() {
+  return <ChatWidget />;
+}
